+++ README.md
@@ -33,3 +33,4 @@ UI:
 
 - Создание скриншота при падении теста
 - Подробное логирование
+- Кроссбраузерное тестирование (Chrome и FireFox)
