+++ src/test/java/api/dashboard/CreateDashboardPositiveAPITest.java
@@ -9,35 +9,20 @@ import io.qameta.allure.Story;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.Test;
 
-public class CreateDashboardAPITest extends BaseAPITest {
+public class CreateDashboardPositiveAPITest extends BaseAPITest {
 
-    Dashboard dashboard;
+    Dashboard dashboard = Dashboard.builder()
+            .name("Test Dashboard-".concat(faker.regexify("[A-Za-z0-9]{8}")))
+            .build();
 
     @Epic("Create Dashboard ")
     @Feature("Dashboard")
     @Story("Create Dashboard with valid value by API")
     @Test(testName = "Create Dashboard with valid value")
     public void createDashboardWithValidValue() {
-        dashboard = Dashboard.builder()
-                .name("Test Dashboard-".concat(faker.regexify("[A-Za-z0-9]{8}")))
-                .build();
         ResponseWrapper response = dashboardAPIStep.createDashboard(dashboard);
         softAssert.assertTrue(dashboardAPIStep.isDashboardCreated(dashboard), "Dashboard isn't created");
         softAssert.assertEquals(response.getStatusCode(), 201);
-        softAssert.assertAll();
-    }
-
-    @Epic("Create Dashboard ")
-    @Feature("Dashboard")
-    @Story("Create Dashboard with invalid value by API")
-    @Test(testName = "Create Dashboard with invalid value")
-    public void createDashboardWithInvalidValue() {
-        dashboard = Dashboard.builder()
-                .name("12")
-                .build();
-        ResponseWrapper response = dashboardAPIStep.createDashboard(dashboard);
-        softAssert.assertFalse(dashboardAPIStep.isDashboardCreated(dashboard), "Dashboard is created");
-        softAssert.assertEquals(response.getStatusCode(), 400);
         softAssert.assertAll();
     }
